Nu met Jpa

+++ src/main/java/com/taskmanager/taskmanager/repository/SubTaakRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubTaakRepository extends JpaRepository<SubTaak, Long> {
+public interface SubTaakRepository extends JpaRepository<SubTaak, Integer> {
 }
